task/PP-15948: setAmount with int type.

+++ payrexx/src/Service/PayrexxApiService.php
@@ -122,7 +122,7 @@ class PayrexxApiService
             $gateway->setPurpose($purpose);
         }
 
-        $gateway->setAmount($total * 100);
+        $gateway->setAmount((int) ($total * 100));
         $gateway->setCurrency($currency);
         $gateway->setSuccessRedirectUrl($redirectUrls['success']);
         $gateway->setCancelRedirectUrl($redirectUrls['cancel']);
